# Make fill_buffer a public API and move the impl to C++