--- a/torchaudio/csrc/ffmpeg/stream_reader/stream_reader_tensor_binding.cpp
+++ b/torchaudio/csrc/ffmpeg/stream_reader/stream_reader_tensor_binding.cpp
@@ -167,6 +167,11 @@ TORCH_LIBRARY_FRAGMENT(torchaudio, m) {
             return s->process_packet(timeout, backoff);
           })
       .def("process_all_packets", [](S s) { s->process_all_packets(); })
+      .def(
+          "fill_buffer",
+          [](S s, const c10::optional<double>& timeout, const double backoff) {
+            return s->fill_buffer(timeout, backoff);
+          })
       .def("is_buffer_ready", [](S s) { return s->is_buffer_ready(); })
       .def("pop_chunks", [](S s) { return s->pop_chunks(); });
 }